+++ jest.screenshots.config.js
@@ -1,7 +1,7 @@
 module.exports = {
     preset: 'ts-jest/presets/js-with-ts',
     testEnvironment: 'node',
-    setupFilesAfterEnv: ['./packages/setupScreenshotsTests.ts'],
+    setupFilesAfterEnv: ['./tools/jest/setupScreenshotsTests.ts'],
     modulePathIgnorePatterns: ['dist'],
     globals: {
         'ts-jest': {
@@ -11,9 +11,11 @@ module.exports = {
                     '@babel/plugin-proposal-nullish-coalescing-operator',
                 ],
             },
+            tsconfig: 'tsconfig.test.json',
         },
     },
     moduleNameMapper: {
+        '@alfalab/core-components-(.*)/(.*)$': '<rootDir>/packages/$1/src/$2',
         '@alfalab/core-components-(.*)$': '<rootDir>/packages/$1/src',
         '\\.css$': 'identity-obj-proxy',
     },