+++ packages/codemod/src/42-calendar/transform.ts
@@ -3,7 +3,8 @@ import {
     hasImportSpecifier,
     isIndexEntry,
     modifyIdentifiers,
-    modifyImportPath, removeEmptyCoreImport,
+    modifyImportPath,
+    removeEmptyCoreImport,
 } from '../42-utils';
 
 const calendarEntryTransformer = (source, j) => {